requirements.txt
PyYAML
python-dotenv
requests
feedparser
beautifulsoup4
openai
pydantic
backoff
notion-client
pytest